various fixes
* layerImages() doesn't block draw()
* layerImages() doesn't delete images
* loadImages() doesn't crash with invalid
* folder path or non picture files in folder
* moved if statements used for sequencing image processing into draw so it was easier for me to understand the flow
* some changes to width used for showing images and use of rightBuffer to better center image
* recolor() temporarily sets color mode to RGB - THIS WAS THE BUG THAT WAS MAKING THE IMAGE BLACK SINCE THE 0 BLUE BECAME 0 BRIGHTNESS

Most of the changes shown in github come from the fact that I'm in the habit of running auto format before running the code and I realized too late. The most major change it made was adding a space after if, and moving "else"s up one line. If you don't want to merge this code that's fine.