Split up loading files and processing images. UI layout updates. Added unprocessed image view in top right corner.
Load files and process images are now separate buttons. For some reason processing images only works when the button is pressed quickly and repeatedly. I'm guessing some kind of multiple threads issue again but I can't figure out what part of the code is causing it. Mini viewer is a WIP, I think I will make a variant of the button class for vertical left/right arrow buttons. The plan is to be able to cycle through the unprocessed images before processing so the user can make sure the selected folder is correct.