Progress bar is now functional
Rectangle around the progress bar feels a little crowded and I'm not sure how to fix that without adjusting all of the positioning. Uncomment line 90 to see what it looks like without the rectangle. Load files button was able to interrupt loading files so I tried disabling it while loadImages was running, but for some reason loadImages only worked while the button was enabled. Uncomment line 182 to test.