Added animation to the overlay strength slider
Used easeValue function to smoothly change the Y coordinate of the overlay strength slider. Also uses easeValue to modify the alpha value I just added to the slider class, so it can become transparent. A rectangle is drawn under the contrast slider (another addition to the Slider class, used by setting backgroundOn to true) to prevent the sliders from appearing to overlap, and providing a "drawer" for the hidden slider to come out of